+++ Clay-cs.Example/Examples/IntroducingClay.cs
@@ -66,8 +66,13 @@ public class IntroducingClay : IDisposable
 		RaylibClay.Fonts[0] = Raylib.LoadFont("resources/Roboto-Regular.ttf");
 		Raylib.SetTextureFilter(RaylibClay.Fonts[0].texture, TextureFilter.TEXTURE_FILTER_BILINEAR);
 
+		Clay.SetDebugModeEnabled(true);
+		
 		while (Raylib.WindowShouldClose() == false)
 		{
+			// NOTE: this is here to force GC errors to show up consistently, You should never have this in your actual code!!!
+			GC.Collect();
+
 			Clay.SetLayoutDimensions(new Clay_Dimensions
 			{
 				width = Raylib.GetScreenWidth(),
@@ -79,9 +84,8 @@ public class IntroducingClay : IDisposable
 
 			Clay.BeginLayout();
 
-			using (Clay.Element(new()
+			using (Clay.Element(Clay.Id(_clayString["OuterContainer"]), new()
 			{
-				id = Clay.Id(_clayString["OuterContainer"]),
 				backgroundColor = new Clay_Color(43, 41, 51),
 				layout = new()
 				{
@@ -92,9 +96,8 @@ public class IntroducingClay : IDisposable
 				}
 			}))
 			{
-				using (Clay.Element(new()
+				using (Clay.Element(Clay.Id(_clayString["HeaderBar"]), new()
 				{
-					id = Clay.Id(_clayString["HeaderBar"]),
 					backgroundColor = _contentBackgroundColor,
 					cornerRadius =  Clay_CornerRadius.All(8),
 					layout = new()
@@ -109,9 +112,8 @@ public class IntroducingClay : IDisposable
 					var fileButtonStr = _clayString["FileButton"];
 					var fileMenuStr = _clayString["FileMenu"];
 
-					using (Clay.Element(new()
+					using (Clay.Element(Clay.Id(fileButtonStr), new()
 					{
-						id = Clay.Id(fileButtonStr),
 						layout = new()
 						{
 							padding = Clay_Padding.HorVer(16, 8)
@@ -120,7 +122,7 @@ public class IntroducingClay : IDisposable
 						cornerRadius = Clay_CornerRadius.All(5),
 					}))
 					{
-						Clay.OpenTextElement("File", new()
+						Clay.TextElement("File", new()
 						{
 							fontSize = 16,
 							textColor = new Clay_Color(255, 255, 255),
@@ -131,9 +133,8 @@ public class IntroducingClay : IDisposable
 
 						if (isMenuVisible)
 						{
-							using (Clay.Element(new()
+							using (Clay.Element(Clay.Id(fileMenuStr), new()
 							{
-								id = Clay.Id(fileMenuStr),
 								floating = new()
 								{
 									attachTo = Clay_FloatingAttachToElement.CLAY_ATTACH_TO_PARENT,
@@ -184,9 +185,8 @@ public class IntroducingClay : IDisposable
 					RenderHeaderButton(_clayString["Upload"]);
 				}
 
-				using (Clay.Element(new()
+				using (Clay.Element(Clay.Id(_clayString["LowerContent"]), new()
 				{
-					id = Clay.Id(_clayString["LowerContent"]),
 					layout = new()
 					{
 						sizing = new Clay_Sizing(Clay_SizingAxis.Grow(), Clay_SizingAxis.Grow()),
@@ -194,9 +194,8 @@ public class IntroducingClay : IDisposable
 					}
 				}))
 				{
-					using (Clay.Element(new()
+					using (Clay.Element(Clay.Id(_clayString["Sidebar"]), new()
 					{
-						id = Clay.Id(_clayString["Sidebar"]),
 						backgroundColor = _contentBackgroundColor,
 						layout = new()
 						{
@@ -226,7 +225,7 @@ public class IntroducingClay : IDisposable
 									cornerRadius = Clay_CornerRadius.All(8),
 								}))
 								{
-									Clay.OpenTextElement(document.Title, new()
+									Clay.TextElement(document.Title, new()
 									{
 										fontSize = 20,
 										textColor = new Clay_Color(255, 255, 255),
@@ -235,7 +234,7 @@ public class IntroducingClay : IDisposable
 							}
 							else
 							{
-								using (var sidebarButton = Clay.Element())
+								using (var sidebarButton = Clay.OpenElement())
 								{
 									sidebarButton.Configure(new()
 									{
@@ -254,7 +253,7 @@ public class IntroducingClay : IDisposable
 											_selectedDocumentIndex = index;
 										}
 									});
-									Clay.OpenTextElement(document.Title, new()
+									Clay.TextElement(document.Title, new()
 									{
 										fontSize = 20,
 										textColor = new Clay_Color(255, 255, 255),
@@ -264,30 +263,32 @@ public class IntroducingClay : IDisposable
 						}
 					}
 
-					using (Clay.Element(new()
+					using (var content = Clay.OpenElement(Clay.Id(_clayString["MainContent"])))
 					{
-						id = Clay.Id(_clayString["MainContent"]),
-						scroll = new()
+						content.Configure(new()
 						{
-							vertical = true
-						},
-						layout = new()
-						{
-							layoutDirection = Clay_LayoutDirection.CLAY_TOP_TO_BOTTOM,
-							childGap = 16,
-							padding = Clay_Padding.All(16),
-							sizing = new Clay_Sizing(Clay_SizingAxis.Grow(), Clay_SizingAxis.Grow())
-						},
-						backgroundColor = _contentBackgroundColor,
-					}))
-					{
+							clip = new()
+							{
+								vertical = true,
+								childOffset = Clay.GetScrollOffset(),
+							},
+							layout = new()
+							{
+								layoutDirection = Clay_LayoutDirection.CLAY_TOP_TO_BOTTOM,
+								childGap = 16,
+								padding = Clay_Padding.All(16),
+								sizing = new Clay_Sizing(Clay_SizingAxis.Grow(), Clay_SizingAxis.Grow())
+							},
+							backgroundColor = _contentBackgroundColor,
+						});
+						
 						var doc = _documents[_selectedDocumentIndex];
-						Clay.OpenTextElement(doc.Title, new()
+						Clay.TextElement(doc.Title, new()
 						{
 							fontSize = 24,
 							textColor = new Clay_Color(255, 255, 255),
 						});
-						Clay.OpenTextElement(doc.Contents, new()
+						Clay.TextElement(doc.Contents, new()
 						{
 							fontSize = 24,
 							textColor = new Clay_Color(255, 255, 255),
@@ -318,7 +319,7 @@ public class IntroducingClay : IDisposable
 			cornerRadius = Clay_CornerRadius.All(5),
 		}))
 		{
-			Clay.OpenTextElement(text, new Clay_TextElementConfig
+			Clay.TextElement(text, new Clay_TextElementConfig
 			{
 				fontSize = 16,
 				textColor = new Clay_Color(255, 255, 255),
@@ -336,7 +337,7 @@ public class IntroducingClay : IDisposable
 			}
 		}))
 		{
-			Clay.OpenTextElement(text, new Clay_TextElementConfig
+			Clay.TextElement(text, new Clay_TextElementConfig
 			{
 				fontSize = 16,
 				textColor = new Clay_Color(255, 255, 255),