+++ Clay.Builder/Program.cs
@@ -68,7 +68,9 @@ Target("Interop", () =>
 			"Clay__Clay_ErrorTypeWrapper",
 			"Clay__Clay_ErrorDataWrapper",
 			"Clay__Clay_ErrorHandlerWrapper",
+			"Clay__Clay_ClipElementConfigWrapper",
 			"Clay__StringArray",
+			"Clay__SuppressUnusedLatchDefinitionVariableWarning",
 		]),
 
 		"--namespace Clay_cs",